Move renderer to pull request package
